Updated to move github class method stuff to githubAPI - in process
